requirements.txt
filelock==3.19.1
fsspec==2025.9.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.1
snakeviz==2.2.2
sympy==1.14.0
tifffile==2025.9.9
torch==2.8.0
tornado==6.5.2
typing_extensions==4.15.0
